login authenticate

+++ app/helpers/application_helper.rb
@@ -6,15 +6,14 @@ module ApplicationHelper
       nav += '<li>' + link_to('Show Users', users_path) + '</li>'
     end
 
-    if @current_user.present?
+    if @current_user.present? && session[:user_id]
       nav += '<li>' + link_to("Log out #{@current_user.email}", login_path, :method => :delete ) + '</li>'
     else
       nav += '<li>' +link_to('Sign up', new_user_path) + '</li>'
-      nav += "<form method='post' action='/login' ><label for='email'>Email: </label><input type='email' name='email'/><label for='password'>Password:</label><input type='password' name='password'/><input type='submit'/>
-      </form>"
-
+       nav += '<li>' + link_to('Log in', login_path) + '</li>'
     end
+
     nav
   end
 end
-# 
+# "<form method='post' action='/login' ><label for='email'>Email: </label><input type='email' name='email'/><label for='password'>Password:</label><input type='password' name='password'/><input type='submit'/> </form>"
